create citybuilder package
