+++ src/main/java/com/canoo/cog/ui/citybuilder/Xform.java
@@ -1,4 +1,4 @@
-package com.canoo.cog.ui;
+package com.canoo.cog.ui.citybuilder;
 
 /*
  * #%L
