+++ src/main/java/com/canoo/cog/ui/citybuilder/model/Building.java
@@ -1,4 +1,4 @@
-package com.canoo.cog.ui.model;
+package com.canoo.cog.ui.citybuilder.model;
 
 
 /*
@@ -21,17 +21,14 @@ package com.canoo.cog.ui.model;
  * #L%
  */
 
-import javafx.scene.paint.Color;
+import com.canoo.cog.ui.citybuilder.model.style.CityStyle;
 import javafx.scene.paint.PhongMaterial;
 import javafx.scene.shape.Box;
 import javafx.beans.property.*;
 
 import javafx.scene.image.Image;
-import java.util.Random;
 
 import com.canoo.cog.sonar.model.Model;
-import com.canoo.cog.ui.model.style.CityStyle;
-import com.canoo.cog.ui.model.style.CityStyle.Style;
 
 public class Building extends AbstractElement<Box> {
 
@@ -57,10 +54,10 @@ public class Building extends AbstractElement<Box> {
 		Image image = null;
 		PhongMaterial phongMaterial = new PhongMaterial();
 
-		if (Style.GOTHAM.name().equals(styleProperty.getValue())) {
+		if (CityStyle.Style.GOTHAM.name().equals(styleProperty.getValue())) {
 			phongMaterial.setDiffuseColor(new Painter().getPaintByLevel(super
 					.getLevel()));
-		} else if (Style.TEST_COVERAGE.name().equals(styleProperty.getValue())) {
+		} else if (CityStyle.Style.TEST_COVERAGE.name().equals(styleProperty.getValue())) {
 			if (model.getCoverage() != null) {
 				try {
 					double coverage = Double.valueOf(model.getCoverage());
@@ -73,14 +70,14 @@ public class Building extends AbstractElement<Box> {
 				phongMaterial.setSpecularColor(new Painter().paint());
 			}
 
-		} else if (Style.TECH_DEBT.name().equals(styleProperty.getValue())) {
+		} else if (CityStyle.Style.TECH_DEBT.name().equals(styleProperty.getValue())) {
 			phongMaterial.setDiffuseColor(new Painter().getPaintForSqualeIndex(model.getSqaleIndex(), model.getTopSqualeIndex()));
-		} else if (Style.TECH_DEBT_PER_LINE.name().equals(styleProperty.getValue())) {
+		} else if (CityStyle.Style.TECH_DEBT_PER_LINE.name().equals(styleProperty.getValue())) {
 			phongMaterial.setDiffuseColor(new Painter()
 			.getPaintForSqaleIndexPerLineOfCode(
 					model.getSqualeIndexPerLineOfCode(),
 					model.getTopSqualeIndexPerLineOfCode()));
-		} else if (Style.COMPLEXITY_PER_LINE.name().equals(styleProperty.getValue())) {
+		} else if (CityStyle.Style.COMPLEXITY_PER_LINE.name().equals(styleProperty.getValue())) {
 			phongMaterial.setDiffuseColor(new Painter()
 					.getPaintForComplexityPerLineOfCode(
 							model.getComplexityPerLineOfCode(),
