+++ src/main/java/com/canoo/cog/ui/citybuilder/model/City.java
@@ -1,4 +1,4 @@
-package com.canoo.cog.ui.model.text;
+package com.canoo.cog.ui.citybuilder.model;
 
 /*
  * #%L
@@ -21,14 +21,19 @@ package com.canoo.cog.ui.model.text;
  */
 
 
-public class Info {
-	private static InfoProperty infoProperty = new InfoProperty();
-	public static void update (String info) {
-		infoProperty.setValue(info);
-	}
-	public static InfoProperty getInfoProperty() {
-		return infoProperty;
-	}
+import javafx.scene.shape.Box;
 
+
+public class City extends Hood {
+
+	public City(double inHeight, double inWidth, double inDepth, String info) {
+		super(inHeight ,inWidth, inDepth, 0, 0, info, 0);
+	}
 	
+	protected Box createShape() {
+		Box box =  new Box();
+		box.setHeight(8);
+		box.setTranslateY(4);
+		return box;
+	}
 }
