+++ src/main/java/com/canoo/cog/ui/citybuilder/model/text/InfoProperty.java
@@ -1,4 +1,4 @@
-package com.canoo.cog.ui.model;
+package com.canoo.cog.ui.citybuilder.model.text;
 
 /*
  * #%L
@@ -21,19 +21,12 @@ package com.canoo.cog.ui.model;
  */
 
 
-import javafx.scene.shape.Box;
+import javafx.beans.property.SimpleStringProperty;
 
+public class InfoProperty extends SimpleStringProperty {
 
-public class City extends Hood {
-
-	public City(double inHeight, double inWidth, double inDepth, String info) {
-		super(inHeight ,inWidth, inDepth, 0, 0, info, 0);
-	}
-	
-	protected Box createShape() {
-		Box box =  new Box();
-		box.setHeight(8);
-		box.setTranslateY(4);
-		return box;
+	@Override
+	public String getName() {
+		return "info";
 	}
 }
