+++ src/main/resources/com/canoo/cog/ui/citybuilder/model/style/tdpl.jpg
